+++ src/interfaces/IPushTo.ts
@@ -1,4 +1,4 @@
-import {IAssignTo} from './IAssignTo';
+import { IAssignTo } from './IAssignTo';
 
 export interface IPushTo extends IAssignTo {
     children?: boolean;