+++ src/interfaces/IReport.ts
@@ -1,5 +1,5 @@
-import {IContextBase} from './IContext';
-import {ActionSnapshot} from '../models';
+import { IContextBase } from './IContext';
+import { ActionSnapshot } from '../models';
 
 export interface IReport {
     context: {