+++ src/plugins/fs/index.ts
@@ -1,11 +1,10 @@
 import {IPlugin} from '../../interfaces';
-import {EncryptActionHandler} from './EncryptActionHandler';
-import {DecryptActionHandler} from './DecryptActionHandler';
 import {WriteToFileActionHandler} from './WriteToFileActionHandler';
 import {MakeDirActionHandler} from './MakeDirActionHandler';
 import {RemovePathActionHandler} from './RemovePathActionHandler';
 import {MovePathActionHandler} from './MovePathActionHandler';
 import {CopyPathActionHandler} from './CopyPathActionHandler';
+import { FindActionHandler } from './FindActionHandler';
 
 const version: string = require('../../../../package.json').version;
 
@@ -15,8 +14,7 @@ module.exports = <IPlugin> {
 
     actionHandlers: [
         new CopyPathActionHandler(),
-        new EncryptActionHandler(),
-        new DecryptActionHandler(),
+        new FindActionHandler(),
         new MakeDirActionHandler(),
         new MovePathActionHandler(),
         new RemovePathActionHandler(),