+++ src/schemas/action.ts
@@ -1,16 +1,18 @@
 import * as Joi from 'joi';
-import {FBLService} from '../services';
+import { FBLService } from '../services';
 
 const joiStepSchemaExt = Joi.extend({
     name: 'FBLAction',
-    base: Joi.object().min(1).required(),
+    base: Joi.object()
+        .min(1)
+        .required(),
     language: {
-        fields: ''
+        fields: '',
     },
     rules: [
         {
             name: 'fields',
-            validate (params, value, state, options) {
+            validate(params, value, state, options) {
                 const keys = Object.keys(value);
 
                 let nonAnnotationKeys = 0;
@@ -21,15 +23,20 @@ const joiStepSchemaExt = Joi.extend({
                 }
 
                 if (nonAnnotationKeys !== 1) {
-                    return this.createError(`Found ${nonAnnotationKeys} non-annotation fields, but only one is required.`, {}, state, options);
+                    return this.createError(
+                        `Found ${nonAnnotationKeys} non-annotation fields, but only one is required.`,
+                        {},
+                        state,
+                        options,
+                    );
                 }
 
                 return value;
-            }
-        }
-    ]
+            },
+        },
+    ],
 });
 
 const FBL_ACTION_SCHEMA = joiStepSchemaExt.FBLAction().fields();
 
-export {FBL_ACTION_SCHEMA};
+export { FBL_ACTION_SCHEMA };