+++ src/schemas/assignTo.ts
@@ -10,21 +10,13 @@ const FBL_ASSIGN_TO_OBJECT_SCHEMA = Joi.object({
     parameters: Joi.string()
         .regex(/^\$\.[^.]+(\.[^.]+)*$/)
         .min(1),
-    override: Joi.boolean()
+    override: Joi.boolean(),
 }).or('ctx', 'secrets', 'parameters');
 
-const FBL_ASSIGN_TO_STRING_SCHEMA = Joi
-    .string()
+const FBL_ASSIGN_TO_STRING_SCHEMA = Joi.string()
     .regex(/^\$\.(ctx|secrets|parameters)\.[^.]+(\.[^.]+)*$/)
     .min(1);
 
-const FBL_ASSIGN_TO_SCHEMA = Joi.alternatives(
-    FBL_ASSIGN_TO_OBJECT_SCHEMA,
-    FBL_ASSIGN_TO_STRING_SCHEMA
-);
+const FBL_ASSIGN_TO_SCHEMA = Joi.alternatives(FBL_ASSIGN_TO_OBJECT_SCHEMA, FBL_ASSIGN_TO_STRING_SCHEMA);
 
-export {
-    FBL_ASSIGN_TO_OBJECT_SCHEMA,
-    FBL_ASSIGN_TO_STRING_SCHEMA,
-    FBL_ASSIGN_TO_SCHEMA
-};
+export { FBL_ASSIGN_TO_OBJECT_SCHEMA, FBL_ASSIGN_TO_STRING_SCHEMA, FBL_ASSIGN_TO_SCHEMA };